backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Fuzzy Matching & NLP
rapidfuzz==3.6.1
jellyfish==1.0.3
unidecode==1.3.8
python-dateutil==2.8.2

# Background Tasks & Scheduling
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# Data Processing
pandas==2.1.4
httpx==0.26.0
aiofiles==23.2.1

# Reporting
openpyxl==3.1.2
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0

